Move visualizers into a seperate file
