+++ src/pymor/gui/qt/widgets/gl.py
