+++ src/pymor/gui/qt/widgets/matplotlib.py
