+++ src/pymor/gui/qt/windows.py
@@ -1,0 +1,155 @@
+try:
+    from PySide.QtGui import (QWidget, QVBoxLayout, QHBoxLayout, QGridLayout, QSlider, QApplication, QLCDNumber,
+                              QAction, QStyle, QToolBar, QLabel, QFileDialog, QMessageBox)
+    from PySide.QtCore import Qt, QCoreApplication, QTimer
+    HAVE_PYSIDE = True
+except ImportError:
+    HAVE_PYSIDE = False
+
+import math as m
+
+if HAVE_PYSIDE:
+
+    class PlotMainWindow(QWidget):
+        """Base class for plot main windows."""
+
+        def __init__(self, U, plot, length=1, title=None):
+            super(PlotMainWindow, self).__init__()
+
+            layout = QVBoxLayout()
+
+            if title:
+                title = QLabel('<b>' + title + '</b>')
+                title.setAlignment(Qt.AlignHCenter)
+                layout.addWidget(title)
+            layout.addWidget(plot)
+
+            plot.set(U, 0)
+
+            if length > 1:
+                hlayout = QHBoxLayout()
+
+                self.slider = QSlider(Qt.Horizontal)
+                self.slider.setMinimum(0)
+                self.slider.setMaximum(length - 1)
+                self.slider.setTickPosition(QSlider.TicksBelow)
+                hlayout.addWidget(self.slider)
+
+                lcd = QLCDNumber(m.ceil(m.log10(length)))
+                lcd.setDecMode()
+                lcd.setSegmentStyle(QLCDNumber.Flat)
+                hlayout.addWidget(lcd)
+
+                layout.addLayout(hlayout)
+
+                hlayout = QHBoxLayout()
+
+                toolbar = QToolBar()
+                self.a_play = QAction(self.style().standardIcon(QStyle.SP_MediaPlay), 'Play', self)
+                self.a_play.setCheckable(True)
+                self.a_rewind = QAction(self.style().standardIcon(QStyle.SP_MediaSeekBackward), 'Rewind', self)
+                self.a_toend = QAction(self.style().standardIcon(QStyle.SP_MediaSeekForward), 'End', self)
+                self.a_step_backward = QAction(self.style().standardIcon(QStyle.SP_MediaSkipBackward),
+                                               'Step Back', self)
+                self.a_step_forward = QAction(self.style().standardIcon(QStyle.SP_MediaSkipForward), 'Step', self)
+                self.a_loop = QAction(self.style().standardIcon(QStyle.SP_BrowserReload), 'Loop', self)
+                self.a_loop.setCheckable(True)
+                toolbar.addAction(self.a_play)
+                toolbar.addAction(self.a_rewind)
+                toolbar.addAction(self.a_toend)
+                toolbar.addAction(self.a_step_backward)
+                toolbar.addAction(self.a_step_forward)
+                toolbar.addAction(self.a_loop)
+                if hasattr(self, 'save'):
+                    self.a_save = QAction(self.style().standardIcon(QStyle.SP_DialogSaveButton), 'Save', self)
+                    toolbar.addAction(self.a_save)
+                    self.a_save.triggered.connect(self.save)
+                hlayout.addWidget(toolbar)
+
+                self.speed = QSlider(Qt.Horizontal)
+                self.speed.setMinimum(0)
+                self.speed.setMaximum(100)
+                hlayout.addWidget(QLabel('Speed:'))
+                hlayout.addWidget(self.speed)
+
+                layout.addLayout(hlayout)
+
+                self.timer = QTimer()
+                self.timer.timeout.connect(self.update_solution)
+
+                self.slider.valueChanged.connect(self.slider_changed)
+                self.slider.valueChanged.connect(lcd.display)
+                self.speed.valueChanged.connect(self.speed_changed)
+                self.a_play.toggled.connect(self.toggle_play)
+                self.a_rewind.triggered.connect(self.rewind)
+                self.a_toend.triggered.connect(self.to_end)
+                self.a_step_forward.triggered.connect(self.step_forward)
+                self.a_step_backward.triggered.connect(self.step_backward)
+
+                self.speed.setValue(50)
+
+            elif hasattr(self, 'save'):
+                hlayout = QHBoxLayout()
+                toolbar = QToolBar()
+                self.a_save = QAction(self.style().standardIcon(QStyle.SP_DialogSaveButton), 'Save', self)
+                toolbar.addAction(self.a_save)
+                hlayout.addWidget(toolbar)
+                layout.addLayout(hlayout)
+                self.a_save.triggered.connect(self.save)
+
+            self.setLayout(layout)
+            self.plot = plot
+            self.U = U
+            self.length = length
+
+        def slider_changed(self, ind):
+            self.plot.set(self.U, ind)
+
+        def speed_changed(self, val):
+            self.timer.setInterval(val * 20)
+
+        def update_solution(self):
+            ind = self.slider.value() + 1
+            if ind >= self.length:
+                if self.a_loop.isChecked():
+                    ind = 0
+                else:
+                    self.a_play.setChecked(False)
+                    return
+            self.slider.setValue(ind)
+
+        def toggle_play(self, checked):
+            if checked:
+                if self.slider.value() + 1 == self.length:
+                    self.slider.setValue(0)
+                self.timer.start()
+            else:
+                self.timer.stop()
+
+        def rewind(self):
+            self.slider.setValue(0)
+
+        def to_end(self):
+            self.a_play.setChecked(False)
+            self.slider.setValue(self.length - 1)
+
+        def step_forward(self):
+            self.a_play.setChecked(False)
+            ind = self.slider.value() + 1
+            if ind == self.length and self.a_loop.isChecked():
+                ind = 0
+            if ind < self.length:
+                self.slider.setValue(ind)
+
+        def step_backward(self):
+            self.a_play.setChecked(False)
+            ind = self.slider.value() - 1
+            if ind == -1 and self.a_loop.isChecked():
+                ind = self.length - 1
+            if ind >= 0:
+                self.slider.setValue(ind)
+
+else:
+
+    class PlotMainWindow(object):
+        pass
